+++ src/main.rs
@@ -1,6 +1,20 @@
 use std::io::IsTerminal;
 
+static USAGE: &str = "\
+Access the text of the Unlicense. <https://unlicense.org>
+
+When used interactively, writes to ./UNLICENSE.
+Otherwise (e.g. in a pipe), writes to stdout.
+
+When given any arguments, displays this help page.
+";
+
 fn main() {
+    if std::env::args().count() > 1 {
+        print!("{}", USAGE);
+        std::process::exit(0);
+    }
+
     if !std::io::stdout().is_terminal() {
         print!("{}", unlicense::TEXT);
         std::process::exit(0);