+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Running interactively now writes to the file `./UNLICENSE`.
   Running non-interactively (e.g. pipe) still writes to stdout.
 
+### Added
+
+- Display help page when given any arguments.
+
 ## [0.1.0] - 2024-04-04
 
 ### Added